fxml для клиенского GUI
