+++ src/ru/avalon/javapp/devj140/userGUI/MainApplication/Common.java
@@ -1,0 +1,8 @@
+package ru.avalon.javapp.devj140.userGUI.MainApplication;
+
+import ru.avalon.javapp.devj140.userGUI.BusinessLogic.SharedResource;
+
+public class Common {
+    public static SharedResource sharedResource = new SharedResource();
+
+}
