+++ src/ru/avalon/javapp/devj140/userGUI/MainApplication/FXMLDomainGUI.fxml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import java.lang.*?>
+<?import java.util.*?>
+<?import javafx.scene.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<AnchorPane xmlns="http://javafx.com/javafx"
+            xmlns:fx="http://javafx.com/fxml"
+            fx:controller="ru.avalon.javapp.devj140.userGUI.MainApplication.FXMLDomainGUIController"
+            prefHeight="400.0" prefWidth="600.0">
+    <TableView>
+
+    </TableView>
+
+</AnchorPane>
