+++ src/ru/avalon/javapp/devj140/userGUI/MainApplication/FXMLDomainGUIController.java
@@ -1,0 +1,4 @@
+package ru.avalon.javapp.devj140.userGUI.MainApplication;
+
+public class FXMLDomainGUIController{
+}
